Create pocket client to handle pocket api
