+++ lib/pocket_to_mail.rb
@@ -1,11 +1,8 @@
 require_relative 'ext/ext.rb'
+require_relative 'pocket_to_mail/pocket_client.rb'
 
 module PocketToMail
   def self.init
-    Pocket.configure do |config|
-      config.consumer_key = ENV['API_KEY']
-    end
-
     options = { :address              => "smtp.gmail.com",
                 :port                 => 587,
                 :domain               => 'smtp.gmail.com',
